(hypr) add an alt+tab bind
This might take some getting used to. Not just because of the different
mod key but because it works in a different way than the alt+tabs that
I'm used to. With this one, every workspace remembers it's own previous
workspace, so you can create chains of any length, and if I change to one that
isn't in the cycle I can just hit alt tab to go back to the cycle that I
already had.